Create a module/subdirectory structure for adapters.
Move the workspace adapter source code into an adapters subdirectory.
Move the default workspace adapter into its own "default.py" source
file that will store defaults for the app, which makes it easier to
avoid circular imports.
